Shuffle all the tests around and delete a ton of crap
