+++ src/Shimmer.Tests/Shimmer.Tests.csproj
@@ -150,6 +150,7 @@
     <Compile Include="Client\BitsManagerTests.cs" />
     <Compile Include="Client\CheckForUpdateTests.cs" />
     <Compile Include="Client\DownloadReleasesTests.cs" />
+    <Compile Include="Client\InstallManagerTests.cs" />
     <Compile Include="Client\UpdateManagerTests.cs" />
     <Compile Include="Core\AppDomainHelperTests.cs" />
     <Compile Include="Core\DeltaPackageTests.cs" />
